+++ app/Filament/Resources/Activities/Pages/ListActivities.php
@@ -3,7 +3,6 @@
 namespace App\Filament\Resources\Activities\Pages;
 
 use App\Filament\Resources\Activities\ActivityResource;
-use Filament\Actions\CreateAction;
 use Filament\Resources\Pages\ListRecords;
 
 class ListActivities extends ListRecords
@@ -13,7 +12,7 @@ class ListActivities extends ListRecords
     protected function getHeaderActions(): array
     {
         return [
-            CreateAction::make(),
+            //
         ];
     }
 }